+++ lib/Payever/Core/Lock/MySQLLock.php
@@ -59,13 +59,9 @@ class MySQLLock implements LockInterface
      */
     public function releaseLock($lockName)
     {
-        $statement = $this->pdo->prepare("SELECT RELEASE_LOCK(?)");
+        $statement = "SELECT RELEASE_LOCK({$this->prepareLockName($lockName)})";
 
-        $statement->execute([
-            $this->prepareLockName($lockName)
-        ]);
-
-        $statement->fetch(\PDO::FETCH_NUM);
+        $this->pdo->exec($statement);
     }
 
     /**
@@ -74,6 +70,18 @@ class MySQLLock implements LockInterface
      */
     private function prepareLockName($lockName)
     {
-        return $this->pdo->quote("{$lockName}");
+        return $this->pdo->quote("{$this->getCurrentDatabase()}.{$lockName}");
+    }
+
+    /**
+     * @return string|null
+     */
+    private function getCurrentDatabase()
+    {
+        if (null === $this->currentDatabase) {
+            $this->currentDatabase = $this->pdo->query('SELECT DATABASE()')->fetchColumn();
+        }
+
+        return $this->currentDatabase;
     }
 }